# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FarmFlow API
  description: |
    Egyptian livestock marketplace API.
    All monetary values are in EGP (Egyptian Pound).
    All dates are ISO 8601 strings.
    Image paths are relative to the server root (e.g. `/uploads/filename.jpg`).

    **Authentication:** Pass a JWT token in the `Authorization: Bearer <token>` header.

    **Roles:**
    - `buyer` — can browse, order, review, manage favorites
    - `seller` — can manage listings, animals, dairy, supplies, finances
    - `admin` — full access including approval workflows

    **⚠️ Mobile app endpoint notes:**
    - `GET /api/listings/my` is a dedicated route (seller/admin JWT required) — returns own listings (all statuses)
    - `GET /api/orders/my` is handled by `GET /api/orders` (buyer JWT → own orders only)
    - `PATCH /api/listings/:id/status` does not exist — use `PUT /api/listings/:id` with a `status` body field
    - `PATCH /api/admin/users/:id` does not exist — use `PATCH /api/admin/users/:id/toggle`
  version: 1.0.0
  contact:
    email: a.khalil@soar.inc
servers:
  - url: http://localhost:5001/api
    description: Local development
  - url: https://api.farmflow.app/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MongoId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 64f1a2b3c4d5e6f7a8b9c0d1
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              msg:   { type: string }
              param: { type: string }
              value: {}

    # ── Auth ────────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        _id:           { $ref: '#/components/schemas/MongoId' }
        name:          { type: string }
        email:         { type: string, format: email }
        phone:         { type: string }
        role:          { type: string, enum: [buyer, seller, admin] }
        isActive:      { type: boolean }
        # Seller-only
        farmName:      { type: string }
        farmPhone:     { type: string }
        personalPhone: { type: string }
        governorate:   { type: string }
        farmDescription: { type: string }
        farmBanner:    { type: string, description: '/uploads/...' }
        farmLocation:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
            address: { type: string }
        farmCertificates:
          type: array
          items: { type: string }
        averageRating:  { type: number }
        reviewCount:    { type: integer }
        fcmToken:       { type: string, nullable: true, description: Firebase Cloud Messaging device token (set by Flutter app) }
        createdAt:      { type: string, format: date-time }
...
    Listing:
      type: object
      properties:
        _id:        { $ref: '#/components/schemas/MongoId' }
        seller:     { $ref: '#/components/schemas/MongoId' }
        type:
          type: string
          enum: [cattle, buffalo, sheep, goat, camel, horse, poultry, rabbit, other]
        breed:      { type: string }
        age:        { type: number, description: Age in months }
        weight:     { type: number, description: Weight in kg }
        price:      { type: number, description: Total price in EGP }
        pricePerKg: { type: number }
        deliveryType:
          type: string
          enum: [none, farm, admin]
        deliveryCost:      { type: number }
        description:       { type: string }
        images:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected, sold]
        rejectionReason:   { type: string }
        eidAvailable:      { type: boolean }
        slaughterService:  { type: boolean }
        slaughterCost:     { type: number }
        qurbaniShares:
          type: array
          items:
            type: object
            properties:
              shareType:   { type: string, enum: [seventh, quarter, half] }
              pricePerShare: { type: number }
              totalShares:   { type: integer }
              bookedShares:  { type: integer }
        createdAt: { type: string, format: date-time }

    # ── Orders ──────────────────────────────────────────────────────────────────
...
    DairyProduct:
      type: object
      properties:
        _id:         { $ref: '#/components/schemas/MongoId' }
        seller:      { $ref: '#/components/schemas/MongoId' }
        name:        { type: string }
        type:        { type: string, enum: [milk, cheese, yogurt, butter, cream, ghee, other] }
        quantity:    { type: number }
        unit:        { type: string, enum: [kg, liter, piece, pack, dozen] }
        pricePerUnit: { type: number }
        availableFrom: { type: string, format: date }
        availableTo:   { type: string, format: date }
        deliveryAvailable: { type: boolean }
        deliveryCost:  { type: number }
        description:   { type: string }
        images:        { type: array, items: { type: string } }
        status:        { type: string, enum: [pending, approved, rejected] }
        rejectionReason: { type: string }
        minStockAlert: { type: number }
        stockLog:
          type: array
          items:
            type: object
            properties:
              change: { type: number }
              note:   { type: string }
              date:   { type: string, format: date-time }
        createdAt: { type: string, format: date-time }

    # ── Supplies ────────────────────────────────────────────────────────────────
...
    EidConfig:
      type: object
      properties:
        eidMode: { type: boolean }
        eidDate: { type: string, format: date }

# ═══════════════════════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /listings:
    get:
      tags: [Listings]
      summary: Get listings (public — approved only; seller JWT → own listings)
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected, sold] }
          description: Admin or seller can filter by status
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
          description: Page number for server-side pagination (requires `limit`). When omitted, returns full array (backwards-compatible).
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
          description: Items per page. When `page` is also provided, enables paginated response format.
      responses:
        '200':
          description: Array of listings (plain array when no `page` param; paginated object when `page` is present)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Listing' }
                  - type: object
                    properties:
                      items:   { type: array, items: { $ref: '#/components/schemas/Listing' } }
                      total:   { type: integer }
                      page:    { type: integer }
                      pages:   { type: integer }
                      hasMore: { type: boolean }

    post:
      tags: [Listings]
      summary: Create listing (seller only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, weight, price]
              properties:
                type:             { type: string, enum: [cattle, buffalo, sheep, goat, camel, horse, poultry, rabbit, other] }
                breed:            { type: string }
                age:              { type: number, description: Age in months }
                weight:           { type: number }
                price:            { type: number }
                pricePerKg:       { type: number }
                deliveryType:     { type: string, enum: [none, farm, admin], default: none }
                deliveryCost:     { type: number }
                description:      { type: string }
                eidAvailable:     { type: boolean }
                slaughterService: { type: boolean }
                slaughterCost:    { type: number }
                qurbaniShares:    { type: string, description: JSON-encoded array of share objects }
                images:           { type: array, items: { type: string, format: binary } }
      responses:
        '201':
          description: Listing created (status = pending — awaiting admin approval)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Listing' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
        '401': { description: Unauthorized }
        '403': { description: Forbidden — not a seller }
...
  /dairy:
    get:
      tags: [Dairy]
      summary: Get approved dairy products (public; seller JWT → own products)
      responses:
        '200':
          description: Array of dairy products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DairyProduct' }

    post:
      tags: [Dairy]
      summary: Create dairy product (seller only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, type, quantity, unit, pricePerUnit]
              properties:
                name:              { type: string }
                type:              { type: string, enum: [milk, cheese, yogurt, butter, cream, ghee, other] }
                quantity:          { type: number }
                unit:              { type: string, enum: [kg, liter, piece, pack, dozen] }
                pricePerUnit:      { type: number }
                availableFrom:     { type: string, format: date }
                availableTo:       { type: string, format: date }
                deliveryAvailable: { type: boolean }
                deliveryCost:      { type: number }
                description:       { type: string }
                minStockAlert:     { type: number }
                images:            { type: array, items: { type: string, format: binary } }
      responses:
        '201':
          description: Dairy product created (status = pending)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DairyProduct' }
  /dairy/my:
    get:
      tags: [Dairy]
      summary: Get seller's own dairy products (all statuses)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of dairy products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DairyProduct' }
  /dairy/admin/all:
    get:
      tags: [Dairy]
      summary: Get all dairy products regardless of status (admin only)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of all dairy products (with seller populated)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DairyProduct' }
  /dairy/{id}:
    get:
      tags: [Dairy]
      summary: Get dairy product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Dairy product detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DairyProduct' }

    put:
      tags: [Dairy]
      summary: Update dairy product (seller owns it or admin)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:              { type: string }
                type:              { type: string, enum: [milk, cheese, yogurt, butter, cream, ghee, other] }
                quantity:          { type: number }
                unit:              { type: string, enum: [kg, liter, piece, pack, dozen] }
                pricePerUnit:      { type: number }
                availableFrom:     { type: string, format: date }
                availableTo:       { type: string, format: date }
                deliveryAvailable: { type: boolean }
                deliveryCost:      { type: number }
                description:       { type: string }
                minStockAlert:     { type: number }
                keepImages:        { type: string, description: JSON array of existing paths to keep }
                images:            { type: array, items: { type: string, format: binary } }
      responses:
        '200':
          description: Updated dairy product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DairyProduct' }

    delete:
      tags: [Dairy]
      summary: Delete dairy product
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '204': { description: Deleted }
  /dairy/{id}/approve:
    post:
      tags: [Dairy]
      summary: Admin approve or reject dairy product
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:          { type: string, enum: [approved, rejected] }
                rejectionReason: { type: string }
      responses:
        '200':
          description: Updated dairy product
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DairyProduct' }
  /dairy/{id}/stock:
    post:
      tags: [Dairy]
      summary: Update dairy product stock quantity (seller only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [change]
              properties:
                change: { type: number, description: Positive to add stock, negative to subtract }
                note:   { type: string }
      responses:
        '200':
          description: Updated dairy product with new quantity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DairyProduct' }
        '400':
          description: Stock cannot go below zero
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── SUPPLIES ─────────────────────────────────────────────────────────────────
...
  /favorites:
    get:
      tags: [Favorites]
      summary: Get buyer's saved farms (buyer only)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of seller profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
  /favorites/{sellerId}:
    post:
      tags: [Favorites]
      summary: Save farm to favorites (buyer only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: sellerId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Added to favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

    delete:
      tags: [Favorites]
      summary: Remove farm from favorites (buyer only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: sellerId
          required: true
          schema: { $ref: '#/components/schemas/MongoId' }
      responses:
        '200':
          description: Removed from favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ── MARKET PRICES ─────────────────────────────────────────────────────────────
...
  /eid/config:
    get:
      tags: [Eid]
      summary: Get Eid mode configuration (public)
      responses:
        '200':
          description: Eid config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EidConfig' }

    patch:
      tags: [Eid]
      summary: Update Eid mode and date (admin only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EidConfig' }
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EidConfig' }
...